+++ errors.go
@@ -14,11 +14,20 @@ var (
 )
 
 type taskNotFoundError struct {
-	taskName string
+	taskName   string
+	didYouMean string
 }
 
 func (err *taskNotFoundError) Error() string {
-	return fmt.Sprintf(`task: Task %q not found`, err.taskName)
+	if err.didYouMean != "" {
+		return fmt.Sprintf(
+			`task: Task %q does not exist. Did you mean %q?`,
+			err.taskName,
+			err.didYouMean,
+		)
+	}
+
+	return fmt.Sprintf(`task: Task %q does not exist`, err.taskName)
 }
 
 type multipleTasksWithAliasError struct {