+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Improvements to Fish shell completion
+  ([#897](https://github.com/go-task/task/pull/897)).
 - Added ability to set a different watch interval by setting
   `interval: '500ms'` or using the `--interval=500ms` flag
   ([#813](https://github.com/go-task/task/issues/813), [#865](https://github.com/go-task/task/pull/865)).