+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Unreleased
 
 - Add ability to set `aliases` for tasks and namespaces ([#268](https://github.com/go-task/task/pull/268), [#340](https://github.com/go-task/task/pull/340), [#879](https://github.com/go-task/task/pull/879)).
+- Improvements to Fish shell completion
+  ([#897](https://github.com/go-task/task/pull/897)).
+- Added ability to set a different watch interval by setting
+  `interval: '500ms'` or using the `--interval=500ms` flag
+  ([#813](https://github.com/go-task/task/issues/813), [#865](https://github.com/go-task/task/pull/865)).
 - Add colored output to `--list`, `--list-all` and `--summary` flags ([#845](https://github.com/go-task/task/pull/845), [#874](https://github.com/go-task/task/pull/874)).
 - Fix unexpected behavior where `label:` was being shown instead of the task
   name on `--list`